Fix getImages() using the wrong parameter
It was using the second parameter (`folder`) to test if the directory was
existing, but folder was the name of the target images folder - which is
not what we wanted at all. `source` was the actual folder to test.

Because of that, the code could only work if presskit was called in a
folder which has an images folder in the current working dir, like
`data`.