Make builder async and rework createThumbnails
We had a problem currently: the thumbOrImage was looking for the thumb
to determine whether a thumb should be used. This was working create if
a previous had been done, which generated the thumbs.

However, on a fresh build, thumbOrImage would be called before the
completion of the thumbnail generation.

Thus, on a fresh build only, the thumbnail wouldn't be used.

By reworking the async model of the generation, we know wait for the end
of the thumbnail generation before rendering the template.